Pushing finished version 0.5.5
- Animated all banners
- Added sound music management system, which allows music to be played, fading it in and out, and changing music.
- Added sound player management, allowing player to play sound effects via already in place events, like OnPlayerDeath, OnPlayerShoot
- Added a Manual IP connection setting, so user can enter an IP address they wish to try and connect to.
- Reduced data being sent in every playerMovmenet client/server send.
Changed transform being sent only as a boolean value, isFacingRight, because only need to know if player is facing right or left in a 2D game. sending 1 bool instead of 4 floats.
Changed position to being sent only as a vector2, because 2D game, all z axis takes place on 0.  Sending 2 floats instead of 3.